wasmfs(wasm): preserve JS llseek offsets
Return JS-facing seek offsets as an f64 instead of truncating them through int. This preserves positions in JavaScript's exact-integer range for both BigInt ABI modes.

Tests:
- python3 test/runner.py core0.test_wasmfs_llseek
- python3 test/runner.py core0.test_wasmfs_llseek_no_bigint
- python3 test/runner.py core0.test_fs_llseek core0.test_fs_llseek_rawfs
- python3 -m py_compile test/test_core.py
- git diff --check

+++ system/lib/wasmfs/js_api.cpp
@@ -254,13 +254,13 @@ int _wasmfs_lchmod(const char* path, mode_t mode) {
     AT_FDCWD, (intptr_t)path, mode, AT_SYMLINK_NOFOLLOW);
 }
 
-int _wasmfs_llseek(int fd, off_t offset, int whence) {
+double _wasmfs_llseek(int fd, off_t offset, int whence) {
   __wasi_filesize_t newOffset;
   int err = __wasi_fd_seek(fd, offset, whence, &newOffset);
   if (err > 0) {
-    return -err;
+    return -static_cast<double>(err);
   }
-  return newOffset;
+  return static_cast<double>(newOffset);
 }
 
 int _wasmfs_rename(const char* oldpath, const char* newpath) {
